requirements.txt
alabaster==0.7.10
amqp==2.2.2
appdirs==1.4.3
argon2-cffi==16.3.0
Babel==2.4.0
billiard==3.5.0.3
celery==4.1.0
cffi==1.10.0
coverage==4.3.4
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-braces==1.11.0
django-cors-headers==2.0.2
django-coverage-plugin==1.5.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-environ==0.4.1
django-extensions==1.7.8
django-filter==1.0.2
django-model-utils==2.6.1
django-redis==4.7.0
django-rest-auth==0.9.1
django-test-plus==1.0.17
djangorestframework==3.6.2
docutils==0.13.1
factory-boy==2.8.1
Faker==0.7.12
flake8==3.3.0
imagesize==0.7.1
ipdb==0.10.2
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
kombu==4.1.0
Markdown==2.6.8
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
pkg-resources==0.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
py==1.4.33
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
pytest-sugar==0.8.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.14.2
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
sqlparse==0.2.3
termcolor==1.1.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.12.1
